fixed setting payload sandbox test
